added styling for LandingPage and AddFriend
