+++ src/components/FriendList.jsx
@@ -4,13 +4,20 @@ export default function FriendList({friends}) {
   console.log(friends)
   const listOfFriends  = friends.map((friend,index)=>(
     <div>
-      <h1>{friend.name}</h1>
-      <h2>{friend.address}</h2>
-      <h2>{friend.email}</h2>
+      <div className="shadow-xl shadow-background w-96 h-36 rounded-lg bg-white my-5">
+        <div className="bg-primary text-gray-100 h-10 rounded-lg flex align-middle justify-center">
+          <h3 className="mt-1 justify-self-center align-self-center font-semibold mb-4">{friend.name}</h3>
+        </div>
+        <div className="my-5">
+          <p className="text-base text-primary font-bold">{friend.email}</p>
+          <p className="my-2">{friend.address}</p>
+
+        </div>
+      </div>
     </div>
   ))
   return (
-    <div>
+    <div className=" flex flex-row flex-wrap justify-center ">
       {listOfFriends}
     </div>
   )
