+++ src/components/LandingPage.jsx
@@ -1,218 +1,38 @@
 import axios from "axios";
-import React, { useState, useContext } from "react";
+import React, { useState} from "react";
 import { useNavigate } from "react-router-dom";
-import { userContext } from "../context";
-import Message from "./Message";
-import Trial from "./Trial";
+import LoginBox from "./LoginBox";
+import SignupBox from "./SignupBox";
 const { REACT_APP_BACKEND } = process.env;
 
 export default function LandingPage({ user }) {
-  //React hook to change to home page on successful login
-  const navigate = useNavigate();
-  // State and setter for login details
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  // State and setter for signup and login message
-  const [message, setMessage] = useState("");
-  const [name, setName] = useState("");
-  const [address, setAddress] = useState("");
-
   const [signUpVisible, setSignUpVisible] = useState(false);
   const [logInVisible, setLogInVisible] = useState(true);
 
-  const signUpAttempt = () => {
-    //data to send to backend
-    const data = {
-      email: email,
-      password: password,
-      name: name,
-      address: address,
-    };
-    axios.post(`${REACT_APP_BACKEND}/signup`, data).then((response) => {
-      if (response.data === "Something went wrong when creating a new user") {
-        setMessage("Something went wrong when creating a new user");
-      }
-      // Inform user if username already exists
-      if (response.data === "user exists") {
-        setMessage("Username taken. Please try a different username.");
-      }
-      // If successful, inform user to login
-      if (response.data === "sign up success") {
-        setMessage("Sign up successful, please login!");
-      }
-    });
-  };
-
-  const loginAttempt = () => {
-    const data = {
-      email: email,
-      password: password,
-    };
-
-    axios.post(`${REACT_APP_BACKEND}/login`, data).then((response) => {
-      console.log(response);
-      // Inform user if they did not key in username or password
-      if (response.data === "details missing") {
-        setMessage("Please enter an email and password");
-      }
-      // If username or password incorrect, inform player
-      if (response.data === "The email or password is incorrect") {
-        setMessage("Invalid login. Please try again.");
-      }
-      // If successful, redirect to home page
-      if (response.data.success === true) {
-        const { userId } = response.data;
-        console.log(" data user", userId);
-        user.userSetter(userId);
-        console.log("state user", user);
-        // On successful login, redirect to home page
-        navigate("/");
-      }
-    });
-  };
-
-  const toggleLogInSignUp = () => {
-    setLogInVisible(!logInVisible);
-    setSignUpVisible(!signUpVisible);
-  };
-
-  if (logInVisible) {
-    return (
-      <div id="landing-background">
-        <div className="loginBox">
-          <p className="logo">
-            <i className="fas fa-utensils" />
-            ZoomZoom
-          </p>
-          <div className="loginSmallBox">
-            <div>
-              <h2>Login</h2>
-              <h2 onClick={toggleLogInSignUp}>Sign Up</h2>
-            </div>
-            <input
-              name="email"
-              id="email"
-              placeholder="Email"
-              onChange={(event) => setEmail(event.target.value)}
-            />
-            <input
-              name="password"
-              id="password"
-              type="password"
-              placeholder="Password"
-              onChange={(event) => setPassword(event.target.value)}
-            />
-          </div>
-          <div className="btn-container">
-            <button
-              className="btn login-btn"
-              type="submit"
-              onClick={loginAttempt}
-            >
-              Login{" "}
-            </button>
-          </div>
-
-          <Message displayMessage={message} />
-        </div>
-      </div>
-    );
+  const toggleLogIn = () =>{
+    setLogInVisible(true)
+    setSignUpVisible(false)
+  }
+  const toggleSignUp = () =>{
+    setLogInVisible(false)
+    setSignUpVisible(true)
   }
 
-  if (signUpVisible) {
-    return (
-      <div id="landing-background">
-        <div className="loginBox">
-          <p className="logo">
-            <i className="fas fa-utensils" />
-            ZoomZoom
-          </p>
-          <div className="loginSmallBox">
-            <div>
-              <h2 onClick={toggleLogInSignUp}>Login</h2>
-              <h2>Sign Up</h2>
+  return(
+    <div className="flex flex-grow bg-primary text-white items-center justify-center">
+      <div className="">
+        <div className="text-9xl font-mono font-bold mb-12">Trace<span className="text-secondary">YourEther</span></div>
+        <div className="flex justify-center">
+          <div className=" shadow-background w-2/4 h-60 rounded-lg bg-white my-5">
+            <div className="flex flex-row justify-center">
+              <h6 className="bg-primary w-2/4 text-3xl p-2 border-l-2 border-t-2 rounded-tl-lg" onClick={toggleLogIn}>Log In</h6>
+              <h6 className="bg-primary w-2/4 text-3xl p-2 border-r-2 border-t-2 rounded-tr-lg" onClick={toggleSignUp}>Sign Up</h6>
             </div>
-            <input
-              name="email"
-              id="email"
-              placeholder="Email"
-              onChange={(event) => setEmail(event.target.value)}
-            />
-            <input
-              name="password"
-              id="password"
-              type="password"
-              placeholder="Password"
-              onChange={(event) => setPassword(event.target.value)}
-            />
-            <input
-              name="name"
-              id="name"
-              placeholder="Name"
-              onChange={(event) => setName(event.target.value)}
-            />
-            <input
-              name="address"
-              id="address"
-              placeholder="Address"
-              onChange={(event) => setAddress(event.target.value)}
-            />
+            {logInVisible && <LoginBox user={user}/>}
+            {signUpVisible && <SignupBox/>}
           </div>
-          <div className="btn-container">
-            <button
-              className="btn sub-btn"
-              type="submit"
-              onClick={signUpAttempt}
-            >
-              Sign Up{" "}
-            </button>
-          </div>
-
-          <Message displayMessage={message} />
         </div>
       </div>
-    );
-  }
-
-  // return (
-  //   <div id="landing-background">
-
-  //       <div className="loginBox">
-  //         <p className="logo">
-  //           <i className="fas fa-utensils" />
-  //           ZoomZoom
-  //         </p>
-  //         <div className="loginSmallBox">
-  //           <input
-  //             name="email"
-  //             id="email"
-  //             placeholder="Email"
-  //             onChange={(event) => setEmail(event.target.value)}
-  //           />
-  //           <input
-  //             name="password"
-  //             id="password"
-  //             type="password"
-  //             placeholder="Password"
-  //             onChange={(event) => setPassword(event.target.value)}
-  //           />
-  //         </div>
-  //         <div className="btn-container">
-  //           <button className="btn sub-btn" type="submit" onClick={signUpAttempt}>
-  //             Sign Up
-  //             {' '}
-  //           </button>
-
-  //           <button className="btn login-btn" type="submit" onClick={loginAttempt}>
-  //             Login
-  //             {' '}
-  //           </button>
-  //           <Trial />
-  //         </div>
-
-  //         <Message displayMessage={message} />
-  //       </div>
-
-  //   </div>
-  // );
+    </div>
+  )
 }
